+++ CHANGELOG.md
@@ -18,6 +18,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Security
 
+## [2.17.0] - 2019-11-04
+### Added
+- Add single facility submission endpoint [#896](https://github.com/open-apparel-registry/open-apparel-registry/pull/896)
+
+### Changed
+- Replace facility history feature switch with feature flag [#881](https://github.com/open-apparel-registry/open-apparel-registry/pull/881)
+- Better accommodate outside contributors [#895](https://github.com/open-apparel-registry/open-apparel-registry/pull/895)
+- Allow superusers to change the Contributor of a Source [#901](https://github.com/open-apparel-registry/open-apparel-registry/pull/901)
+
+### Deprecated
+
+### Removed
+- Remove legacy API [#888](https://github.com/open-apparel-registry/open-apparel-registry/pull/888)
+
+### Fixed
+- Handle cases when Source contributor or facility_list are None [#903](https://github.com/open-apparel-registry/open-apparel-registry/pull/903)
+
+### Security
+
 ## [2.16.0] - 2019-10-24
 ### Changed
 - Implement Source model step 3 [#858](https://github.com/open-apparel-registry/open-apparel-registry/pull/858)
@@ -288,7 +307,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/open-apparel-registry/open-apparel-registry/compare/2.16.0...HEAD
+[Unreleased]: https://github.com/open-apparel-registry/open-apparel-registry/compare/2.17.0...HEAD
+[2.17.0]: https://github.com/open-apparel-registry/open-apparel-registry/releases/tag/2.17.0
 [2.16.0]: https://github.com/open-apparel-registry/open-apparel-registry/releases/tag/2.16.0
 [2.15.0]: https://github.com/open-apparel-registry/open-apparel-registry/releases/tag/2.15.0
 [2.14.0]: https://github.com/open-apparel-registry/open-apparel-registry/releases/tag/2.14.0